requirements_asgi.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
boto3==1.28.68
botocore==1.31.68
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.6
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cron-descriptor==1.4.3
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
django-appconf==1.0.6
django-background-tasks==1.2.5
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.5.1
django-compat==1.0.15
django-compressor==4.4
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-haystack==3.2.1
django-import-export==3.2.0
django-js-asset==2.0.0
django-mobile==0.7.0
django-mobile-detect==0.1
django-modeltranslation==0.18.11
django-push-notifications==3.0.2
django-rest-framework==0.1.0
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
EasyProcess==1.1
elastic-transport==8.4.1
elasticsearch==8.10.0
elasticsearch-dsl==8.9.0
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.12.2
firebase-admin==6.5.0
Flask==3.0.0
fonttools==4.53.0
frozenlist==1.4.1
fuzzywuzzy==0.18.0
geographiclib==2.0
geoip2==4.8.0
geopy==2.4.1
gevent==22.10.2
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==2.0.2
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==7.0.0
importlib_resources==6.4.0
incremental==22.10.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.3.6
Levenshtein==0.23.0
libretranslatepy==2.1.1
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.9.0
maxminddb==2.5.2
meta==1.0.2
MouseInfo==0.1.3
msgpack==1.0.8
mss==9.0.1
multidict==6.0.5
mysqlclient==2.1.1
networkx==3.2.1
numpy==1.26.2
odfpy==1.4.1
opencv-python==4.8.1.78
openpyxl==3.1.2
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
paytmchecksum==1.7.0
Pillow==9.5.0
platformdirs==3.8.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.17
PyGetWindow==0.0.9
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.8.0
PyMsgBox==1.0.9
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==3.8.0
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-barcode==0.15.1
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gettext==5.0
python-Levenshtein==0.23.0
pytweening==1.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
PyYAML==6.0
qrcode==7.4.2
rapidfuzz==3.4.0
razorpay==1.3.0
rcssmin==1.1.1
redis==4.6.0
reportlab==3.6.12
requests==2.28.2
rjsmin==1.2.1
rsa==4.9
s3transfer==0.7.0
schedule==1.2.2
scipy==1.13.1
selenium==4.20.0
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tablib==3.4.0
tinycss2==1.2.1
translate==3.6.1
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.15
vine==5.1.0
virtualenv==20.23.1
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
webkit2png==0.8.2
Werkzeug==3.0.1
whitenoise==6.5.0
wikipedia==1.4.0
wsproto==1.2.0
xhtml2pdf==0.2.9
xlrd==2.0.1
XlsxWriter==3.1.0
xlwt==1.3.0
xmltodict==0.13.0
yarl==1.9.4
zipp==3.17.0
zope.event==4.6
zope.interface==6.0
